refactor out quote embed helpers from QuoteCommand
This is so they can be used in the new search command
